bodyprog: rodata migrate `bodyprog_800314EC`, add `globals` split
Some of the rodata between `800314EC` funcs was for `g_GameWorkConst` /
`g_SavegamePtr` etc, didn't make sense to have those being inited part
way through a .c, so split those into globals.c file

(looks like there's a bunch of stage/map data between globals.c and the
next .c too, haven't tried to add that in yet though)

Noticed there's sometimes weird unused bytes in .rodata too, even
bodyprog itself starts with 4 unused bytes at 80024B60, wonder if those
were something compiler inserted between rodata of different files,
maybe can help indicate rodata splits for us
