+++ configs/bodyprog.yaml
@@ -61,7 +61,10 @@ segments:
       - [0x0, rodata]
       - [0x4, .rodata, memcard]
       - [0x138, rodata]
-      - [0x818, .rodata, bodyprog_800314EC] # Jumptable: `SysState_Gameplay_Update`
+      - [0x144, .rodata, bodyprog_800314EC]
+      - [0x1E4, .rodata, globals]
+      - [0x1F8, rodata] # Stage/map data?
+      - [0x618, .rodata, bodyprog_80032D1C] # Jumptable: `SysState_Gameplay_Update`
                                             # Jumptable: `func_8003943C`
                                             # Jumptable & Data: `SysState_GameOver_Update`
       - [0x920, rodata]
@@ -107,6 +110,7 @@ segments:
 
       - [0x9ad0,    c, memcard]
       - [0xC98C,    c, bodyprog_800314EC]
+      - [0xE1BC,    c, bodyprog_80032D1C]
       - [0x1b4ec,   c, view/vc_util]
       - [0x1bf04,   c, bodyprog_80040A64]
       - [0x20F1C,   c, libsd/bodyprog_80045A7C]
